eliminar todas las series

+++ commands/sofi/eliminarSerie.js
@@ -1,9 +1,16 @@
-const { SlashCommandBuilder } = require("discord.js");
+const {
+    SlashCommandBuilder,
+    ActionRowBuilder,
+    ButtonBuilder,
+    ButtonStyle,
+    EmbedBuilder,
+    ComponentType,
+} = require("discord.js");
 const SeriesUsersDB = require("../../schemas/seriesUsersDB");
 
 module.exports = {
     cooldown: 2,
-    uso: " `<nombre serie>`",
+    uso: " `<nombre serie | todo>`",
     data: new SlashCommandBuilder()
         .setName("eliminarserie")
         .setDescription(
@@ -19,18 +26,75 @@ module.exports = {
         const serie = interaction.options.getString("serie");
         const userId = interaction.user.id;
 
-        const cant = await SeriesUsersDB.destroy({
-            where: { userId: userId, serie: serie },
-        });
+        if (serie.toLowerCase() === "todo") {
+            const ok = new ButtonBuilder()
+                .setCustomId("ok")
+                .setLabel("Si")
+                .setStyle(ButtonStyle.Success);
+            const cancel = new ButtonBuilder()
+                .setCustomId("cancel")
+                .setLabel("No")
+                .setStyle(ButtonStyle.Danger);
+            const row = new ActionRowBuilder().setComponents(ok, cancel);
 
-        if (cant > 0) {
-            interaction.reply({
-                content: `<@${userId}> Has eliminado **${serie}** de tu lista de series!`,
+            const pregunta = new EmbedBuilder()
+                .setTitle("⚠️️ Eliminar todas las series ⚠️️")
+                .setDescription(
+                    "Estás seguro de que desas eliminar todas las series de tu lista? Esta acción no puede deshacerse!"
+                )
+                .setColor("#ff0000");
+
+            const cancelado = new EmbedBuilder()
+                .setTitle("⚠️ Eliminar todas las series ⚠️")
+                .setDescription("Cancelado! Tus series siguen como siempre :)")
+                .setColor("#ff00ff");
+
+            const resMsg = await interaction.reply({
+                embeds: [pregunta],
+                components: [row],
+            });
+
+            const collector = resMsg.createMessageComponentCollector({
+                componentType: ComponentType.Button,
+                time: 30_000,
+            });
+
+            collector.on("collect", async (i) => {
+                if (i.customId === "cancel") {
+                    await i.update({
+                        embeds: [cancelado],
+                        components: [],
+                    });
+                } else if (i.customId === "ok") {
+                    const cant = await SeriesUsersDB.destroy({
+                        where: { userId: userId },
+                    });
+
+                    const exito = new EmbedBuilder()
+                        .setTitle("⚠️ Eliminar todas las series ⚠️")
+                        .setDescription(`Se han eliminado **\`${cant}\` series** de tu lista!`)
+                        .setColor("#00ff00");
+
+                    await i.update({
+                        embeds: [exito],
+                        components: [],
+                    });
+                }
             });
         } else {
-            interaction.reply({
-                content: `<@${userId}> No se ha encontrado la serie **${serie}** en tu lista, usa el comando \`/listaseries\` para ver las que coleccionas!`,
+            const cant = await SeriesUsersDB.destroy({
+                where: { userId: userId, serie: serie },
             });
+
+            if (cant > 0) {
+                interaction.reply({
+                    content: `<@${userId}> Has eliminado **${serie}** de tu lista de series!`,
+                });
+            } else {
+                interaction.reply({
+                    content: `<@${userId}> No se ha encontrado la serie **${serie}** en tu lista, usa el comando \`/listaseries\` para ver las que coleccionas!`,
+                });
+            }
         }
     },
 };
